Changed meal_quantity datatype

+++ database/migrations/2019_10_08_182302_create_meals_table.php
@@ -19,7 +19,7 @@ class CreateMealsTable extends Migration
             $table->foreign('user_id')->references('id')->on('users');
             $table->string('meal_name');
             $table->string('meal_time');
-            $table->unsignedBigInteger('quantity');
+            $table->float('quantity');
             $table->date('day');
             $table->float('calories');
             $table->timestamps();
